Refactor old model use names
