+++ src/main/java/seedu/address/experimental/storage/JsonAdaptedStats.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.entity.Stats;
+import seedu.address.obsolete.obsoletemodel.entity.Stats;
 /***/
 public class JsonAdaptedStats {
