+++ src/main/java/seedu/address/experimental/storage/JsonAdaptedTag.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.tag.Tag;
+import seedu.address.obsolete.obsoletemodel.tag.Tag;
 
 /**
  * Jackson-friendly version of {@link Tag}.
