+++ src/main/java/seedu/address/logic/Logic.java
@@ -8,7 +8,8 @@ import seedu.address.experimental.model.ReadOnlyReroll;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.entity.Entity;
+import seedu.address.obsolete.obsoletemodel.Model;
+import seedu.address.obsolete.obsoletemodel.entity.Entity;
 
 /**
  * API of the Logic component
@@ -28,7 +29,7 @@ public interface Logic {
     /**
      * Returns the AddressBook.
      *
-     * @see seedu.address.model.Model#getAddressBook()
+     * @see Model#getAddressBook()
      */
     ReadOnlyReroll getReroll();
