+++ src/main/java/seedu/address/logic/commands/AddEntityCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.experimental.model.Model;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.entity.Entity;
+import seedu.address.obsolete.obsoletemodel.entity.Entity;
 
 /**
  * Adds a person to the address book.
