+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.experimental.model.Model;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.entity.Entity;
+import seedu.address.obsolete.obsoletemodel.entity.Entity;
 
 /**
  * Deletes a person identified using it's displayed index from the address book.
