+++ src/main/java/seedu/address/logic/commands/EditValueCommand.java
@@ -16,13 +16,13 @@ import java.util.Set;
 import seedu.address.commons.util.CollectionUtil;
 import seedu.address.experimental.model.Model;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.entity.Character;
-import seedu.address.model.entity.Entity;
-import seedu.address.model.entity.Item;
-import seedu.address.model.entity.Mob;
-import seedu.address.model.entity.Name;
-import seedu.address.model.entity.Stats;
-import seedu.address.model.tag.Tag;
+import seedu.address.obsolete.obsoletemodel.entity.Character;
+import seedu.address.obsolete.obsoletemodel.entity.Entity;
+import seedu.address.obsolete.obsoletemodel.entity.Item;
+import seedu.address.obsolete.obsoletemodel.entity.Mob;
+import seedu.address.obsolete.obsoletemodel.entity.Name;
+import seedu.address.obsolete.obsoletemodel.entity.Stats;
+import seedu.address.obsolete.obsoletemodel.tag.Tag;
 
 /**
  * Edits the details of an existing person in the address book.
