+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.experimental.model.Model;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.obsolete.obsoletemodel.person.NameContainsKeywordsPredicate;
 
 /**
  * Finds and lists all persons in address book whose name contains any of the argument keywords. Keyword matching is
