+++ src/main/java/seedu/address/logic/parser/EditModeParser.java
@@ -21,13 +21,13 @@ import seedu.address.logic.commands.EditValueCommand.EditItemDescriptor;
 import seedu.address.logic.commands.EditValueCommand.EditMobDescriptor;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.entity.Character;
-import seedu.address.model.entity.Entity;
-import seedu.address.model.entity.Item;
-import seedu.address.model.entity.Mob;
-import seedu.address.model.entity.Name;
-import seedu.address.model.entity.Stats;
-import seedu.address.model.tag.Tag;
+import seedu.address.obsolete.obsoletemodel.entity.Character;
+import seedu.address.obsolete.obsoletemodel.entity.Entity;
+import seedu.address.obsolete.obsoletemodel.entity.Item;
+import seedu.address.obsolete.obsoletemodel.entity.Mob;
+import seedu.address.obsolete.obsoletemodel.entity.Name;
+import seedu.address.obsolete.obsoletemodel.entity.Stats;
+import seedu.address.obsolete.obsoletemodel.tag.Tag;
 
 /**
  * Parses user input while in edit mode.
