+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -9,12 +9,12 @@ import java.util.Set;
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.entity.Classification;
-import seedu.address.model.entity.Name;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Phone;
-import seedu.address.model.tag.Tag;
+import seedu.address.obsolete.obsoletemodel.entity.Classification;
+import seedu.address.obsolete.obsoletemodel.entity.Name;
+import seedu.address.obsolete.obsoletemodel.person.Address;
+import seedu.address.obsolete.obsoletemodel.person.Email;
+import seedu.address.obsolete.obsoletemodel.person.Phone;
+import seedu.address.obsolete.obsoletemodel.tag.Tag;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
